+++ lib/augmented/strings.rb
@@ -1,9 +1,11 @@
 require 'augmented/strings/blank'
+require 'augmented/strings/squish'
 require 'augmented/strings/truncatable'
 
 module Augmented
   module Strings
     include Blank
+    include Squish
     include Truncatable
   end
 end